requirements/dev.txt
-r base.txt
coverage==4.5.1
flake8==3.5.0
mccabe==0.6.1
pycodestyle==2.3.1
pyflakes==1.6.0
fabric3==1.14.post1

ansible==2.4.3.0
awscli==1.14.64
awslogs==0.10.0

django-debug-toolbar==1.9.1
  sqlparse==0.2.4
factory-boy==2.11.1
  Faker==0.8.15
  text-unidecode==1.2
